% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{cat_eyes}
\alias{cat_eyes}
\title{Draw cat eyes}
\usage{
cat_eyes(x = NULL, size = 1, fill = "black", colour = "black", color = colour)
}
\arguments{
\item{x}{A portrait object.}

\item{size}{The size of the inner eye outline.}

\item{fill}{The color of the eye.}

\item{colour}{The color of the inner eye outline.}
}
\description{
Draw cat eyes
}
