% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyes.R
\name{eyesGrob}
\alias{eyesGrob}
\title{Eyes grob}
\usage{
eyesGrob(
  x = 0.5,
  y = 0.5,
  size = 0.05,
  dist = 0.4,
  size.outline = 1,
  colour = "black",
  color = colour,
  fill = "black",
  vp = NULL
)
}
\arguments{
\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{size}{size of the pupils}

\item{dist}{the pupilary distance}

\item{size.outline}{the size of the eye outline}

\item{colour, color}{the color of the eye outline}

\item{fill}{the color of the eye}

\item{vp}{viewport}
}
\description{
Eyes grob
}
