% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouth.R
\name{mouthGrob}
\alias{mouthGrob}
\title{Mouth grob}
\usage{
mouthGrob(
  x = 0.5,
  y = 0.5,
  size = 1,
  colour = NULL,
  color = colour,
  vp = NULL,
  smile = 1
)
}
\arguments{
\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{size}{size of the pupils}

\item{vp}{viewport}
}
\description{
Mouth grob
}
