% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shapeGrob}
\alias{shapeGrob}
\title{Draw the shape of the face}
\usage{
shapeGrob(
  x = 0.5,
  y = 0.5,
  width = 1,
  height = 1,
  colour = "black",
  color = colour,
  fill = NA,
  size = 1
)
}
\arguments{
\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{width}{the width of the face}

\item{height}{the height of the face}

\item{colour, color}{the color of the face outline}

\item{fill}{the color of the face}

\item{size}{the size of the outer face shape}
}
\description{
Only supports the round face at the moment.
}
