% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portrait.R
\name{face}
\alias{face}
\title{Draw a standard face}
\usage{
face(shape = "round", eyes = "round", mouth = "smile", ...)
}
\arguments{
\item{shape}{Currently only supports "round" face.}

\item{eyes}{Currently only supports "round" eyes}

\item{mouth}{Currently supports "smile" and "sad".}

\item{...}{Not used currently.}
}
\description{
Draw a standard face
}
